fix: fully validate the OG image's first node before rendering
opengraph-image.tsx only checked first.verse was truthy, not that it
had the fields renderOgCard/clampBody actually use — a stored node
with a malformed verse (e.g. {"ref":"2:255"}, missing translation)
still crashed on clampBody(undefined).trim(). Extracted the shape
check already used by POST /api/share's validation (ref, surahName,
translation as strings) into lib/share-canvas.ts and reused it here,
so both entry points agree on what a valid node looks like.
