+++ app/api/share/[id]/opengraph-image.tsx
@@ -3,6 +3,7 @@ import { db } from "@/lib/db";
 import { sharedCanvases } from "@/lib/db/schema";
 import { eq } from "drizzle-orm";
 import { renderOgCard, clampBody, OG_SIZE, OG_CONTENT_TYPE } from "@/lib/og-card";
+import { isValidNode } from "@/lib/share-canvas";
 import type { SavedCanvas } from "@/store/canvas";
 
 export const alt = "Shared canvas — Open Hikmah";
@@ -46,7 +47,7 @@ export default async function Image({ params }: { params: Promise<{ id: string }
 
   const first = canvas.nodes[0];
   const count = canvas.nodes.length;
-  if (!first?.verse) return fallback();
+  if (!isValidNode(first)) return fallback();
 
   return new ImageResponse(
     renderOgCard({
